dev: wire-up `.gitignore`
Refactor reach command into its own file.
