+++ cspell.json
@@ -45,6 +45,7 @@
         "test-fixtures/**",
         "test-packages/test-cspell-tools/src/*.txt"
     ],
+    "useGitignore": true,
     "ignoreWords": [
         "commitcomment"
     ],
