+++ packages/cspell/package.json
@@ -74,6 +74,7 @@
     "comment-json": "^4.1.1",
     "cspell-glob": "^5.11.1",
     "cspell-lib": "^5.11.1",
+    "cspell-gitignore": "^5.11.1",
     "fast-json-stable-stringify": "^2.1.0",
     "file-entry-cache": "^6.0.1",
     "fs-extra": "^10.0.0",
