+++ packages/cspell/samples/.cspell.json
@@ -14,12 +14,7 @@
     ],
     "maxNumberOfProblems": 10000,
     "ignorePaths": [
-        "forbidden-words.txt",
-        "dictionaries/**",
-        "node_modules/**",
-        "vscode-extension/**",
-        ".git/**",
-        ".vscode/**"
+        "forbidden-words.txt"
     ],
     // flagWords - list of words to be always considered incorrect
     // This is useful for offensive words and common spelling errors.
