+++ packages/cspell/samples/.gitignore
@@ -1,0 +1,9 @@
+# Local globs to ignore
+
+ignore/**
+!ignore/keepme.md
+dictionaries/**
+node_modules/**
+vscode-extension/**
+.git/**
+.vscode/**
